refactor: move input schemas to dedicated schemas folder
