+++ backend/src/schemas/postSchemas.ts
@@ -1,0 +1,17 @@
+import Joi from "joi";
+
+export const createPostSchema = Joi.object({
+  title: Joi.string().min(1).max(255).required(),
+  content: Joi.string().min(1).max(50000).required(),
+  excerpt: Joi.string().max(500).optional(),
+  imageUrl: Joi.string().uri().optional(),
+  tags: Joi.array().items(Joi.string().max(50)).max(10).optional(),
+});
+
+export const updatePostSchema = Joi.object({
+  title: Joi.string().min(1).max(255).optional(),
+  content: Joi.string().min(1).max(50000).optional(),
+  excerpt: Joi.string().max(500).optional(),
+  imageUrl: Joi.string().uri().optional(),
+  tags: Joi.array().items(Joi.string().max(50)).max(10).optional(),
+});
